requirements.txt
boto3==1.34.69
streamlit==1.32.2
python-dotenv==1.0.1
botocore==1.34.69